+++ gauguin-app/src/main/kotlin/org/piepmeyer/gauguin/ui/grid/GridCellSizeListener.kt
@@ -1,5 +1,5 @@
 package org.piepmeyer.gauguin.ui.grid
 
 fun interface GridCellSizeListener {
-    fun cellSizeChanged(cellSizePercent: Int)
+    fun cellSizeChanged()
 }