+++ gauguin-app/src/main/kotlin/org/piepmeyer/gauguin/ui/grid/GridCellSizeService.kt
@@ -1,12 +1,6 @@
 package org.piepmeyer.gauguin.ui.grid
 
 class GridCellSizeService {
-    var cellSizePercent = 100
-        set(cellSizePercent) {
-            field = cellSizePercent
-            listener?.cellSizeChanged(cellSizePercent)
-        }
-
     private var listener: GridCellSizeListener? = null
 
     fun setCellSizeListener(listener: GridCellSizeListener?) {