+++ gauguin-app/src/main/kotlin/org/piepmeyer/gauguin/ui/grid/GridUiInjectionStrategy.kt
@@ -15,4 +15,6 @@ interface GridUiInjectionStrategy {
     fun showOperators(): Boolean
 
     fun markDuplicatedInRowOrColumn(): Boolean
+
+    fun maximumCellSizeInDP(): Int
 }